Refactor code
Move UI methods to new UI controller
