+++ src/main/resources/com.elm/fxml/elm.fxml
@@ -1,4 +1,3 @@
-<?import javafx.geometry.Insets?>
 <?import javafx.scene.layout.GridPane?>
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.Label?>
@@ -6,8 +5,7 @@
 <?import javafx.scene.control.TextField?>
 <?import javafx.scene.control.TextArea?>
 
-
-<GridPane fx:controller="com.elm.controller.MessageController"
+<GridPane fx:controller="com.elm.controller.UIController"
           xmlns:fx="http://javafx.com/fxml" alignment="center" hgap="10" vgap="10">
     <children>
